backend/requirements.txt
django>=5.2.6
django-environ>=0.12.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0
pillow>=11.3.0
psycopg2-binary-2.9.10

